DONE OMG. The solution was that bash can call TCP
So yeah. After being verrry mean to the code (the code now crashes
predictably if the executed shell command didn't return a 0 status
code). This helped in the exploration, as we were now able to figure out
what packages are actually installed - resulting in the discovery of
bash and the subsequent usage of bashs tcp feature to expose the flag.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>